requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.4
flake8==6.0.0
flower==2.0.0
gunicorn==21.2.0
humanize==4.7.0
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kombu==5.3.1
mccabe==0.7.0
packaging==23.1
Pillow==10.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.0
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
tornado==6.3.2
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
